refactor(errors): move error handlers to a concern
refactor(errors): move error handlers to a concern
