+++ backend/app/controllers/api/github_events_controller.rb
@@ -6,10 +6,9 @@ module Api
   # GithubEventsController handles incoming GitHub webhook events.
   # It verifies the GitHub webhook signature and creates a GithubEvent record.
   class GithubEventsController < ApplicationController
-    before_action :verify_signature, only: :create
+    include ErrorHandler
 
-    rescue_from StandardError, with: :handle_standard_error
-    rescue_from ActiveRecord::RecordInvalid, with: :handle_record_invalid
+    before_action :verify_signature, only: :create
 
     # Handles incoming GitHub event
     #
@@ -32,8 +31,8 @@ module Api
     def verify_signature
       return if GitHubSignatureValidator.call request
 
-      render_error 'Invalid GitHub webhook signature',
-                   :forbidden
+      render json: { error: 'Invalid GitHub webhook signature' },
+             status: :forbidden
     end
 
     def event_params
@@ -44,24 +43,6 @@ module Api
         repo_name: payload.dig('repository', 'name'),
         payload:
       }
-    end
-
-    # Handles ActiveRecord::RecordInvalid exceptions
-    def handle_record_invalid(error)
-      render_error "Failed to record event: #{error.message}",
-                   :unprocessable_entity
-    end
-
-    # Handles StandardError exceptions
-    def handle_standard_error(error)
-      render_error "Unexpected error: #{error.message}",
-                   :internal_server_error
-    end
-
-    # Renders error messages
-    def render_error(message, status)
-      render json: { error: message },
-             status:
     end
   end
 end
